+++ feasto_core_clean_arch/interactors/delete_restaurant_interactor.py
@@ -1,4 +1,4 @@
-from feasto_core_clean_arch.exceptions.custom_exceptions import InvalidRestaurantId
+from feasto_core_clean_arch.exceptions.custom_exceptions import InvalidRestaurantId, InvalidRestaurantOwnerId
 from feasto_core_clean_arch.interactors.storage_interfaces.storage_interface import StorageInterface
 from feasto_core_clean_arch.presenters.presenter_implementation import PresenterImplementation
 
@@ -13,10 +13,14 @@ class DeleteRestaurantInteractor:
                  presenter: PresenterImplementation,
                  user_id: int
                  ) :
+        restaurant_owner_id = self.storage.get_restaurant(restaurant_id=restaurant_id).id
+        if restaurant_owner_id != user_id:
+            raise InvalidRestaurantOwnerId
+
         try:
             self.storage.validate_restaurant_id(restaurant_id=restaurant_id)
         except InvalidRestaurantId:
-            presenter.error_response_for_restaurant_not_found()
+            presenter.get_error_response_for_restaurant_not_found()
             return
 
         restaurant_dto = self.storage.delete_restaurant(
